Move most account creation code out of account page component
